requirements.txt
# Core LangChain and LangGraph dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# Async support
aiohttp>=3.9.0

# Type hints and validation
pydantic>=2.0.0
typing-extensions>=4.0.0

# Utilities
python-dotenv>=1.0.0

# Data analysis for A/B testing validation
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
